# src/backend/tests/contract/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "ContentTypeParam": {
        "description": "Content type parameter for API.",
        "enum": [
          "ARTICLE",
          "VIDEO",
          "REEL"
        ],
        "title": "ContentTypeParam",
        "type": "string"
      },
...
      "DataDeletionResponse": {
        "description": "Response for the data deletion endpoint.",
        "properties": {
          "deleted": {
            "additionalProperties": true,
            "title": "Deleted",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "deleted",
          "message"
        ],
        "title": "DataDeletionResponse",
        "type": "object"
      },
...
      "EventTypeParam": {
        "description": "Event type parameter for API.",
        "enum": [
          "VIEW_10S",
          "OPEN_SOURCE",
          "SHARE",
          "SAVE",
          "CHAT_START",
          "CHAT_MESSAGE"
        ],
        "title": "EventTypeParam",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InteractionRequest": {
        "description": "Request body for recording an interaction.",
        "properties": {
          "content_item_id": {
            "title": "Content Item Id",
            "type": "integer"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventTypeParam"
          },
          "extra_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Data"
          }
        },
        "required": [
          "content_item_id",
          "event_type"
        ],
        "title": "InteractionRequest",
        "type": "object"
      },
      "InteractionResponse": {
        "description": "Response for interaction recording.",
        "properties": {
          "event_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "InteractionResponse",
        "type": "object"
      },
...
      "PlaylistItem": {
        "description": "Single item in playlist response.",
        "properties": {
          "cluster_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "entities": {
            "items": {
              "type": "string"
            },
            "title": "Entities",
            "type": "array"
          },
          "global_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Global Score"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "title": "Topics",
            "type": "array"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url"
          }
        },
        "required": [
          "id",
          "type",
          "source",
          "source_url",
          "title",
          "description",
          "summary",
          "image_url",
          "video_url",
          "duration",
          "topics",
          "entities",
          "published_at",
          "global_score",
          "cluster_id"
        ],
        "title": "PlaylistItem",
        "type": "object"
      },
      "PlaylistResponse": {
        "description": "Response for playlist endpoint.",
        "properties": {
          "cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlaylistItem"
            },
            "title": "Items",
            "type": "array"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "total_items": {
            "title": "Total Items",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "session_id",
          "cursor",
          "has_more",
          "total_items"
        ],
        "title": "PlaylistResponse",
        "type": "object"
      },
      "PreferencesResponse": {
        "description": "Response for preferences endpoint.",
        "properties": {
          "entities": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Entities",
            "type": "array"
          },
          "formats": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Formats",
            "type": "array"
          },
          "sources": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Sources",
            "type": "array"
          },
          "topics": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Topics",
            "type": "array"
          }
        },
        "required": [
          "topics",
          "entities",
          "sources",
          "formats"
        ],
        "title": "PreferencesResponse",
        "type": "object"
      },
...
      "UserStatsResponse": {
        "description": "Response for user stats endpoint.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "engagement_7d": {
            "additionalProperties": true,
            "title": "Engagement 7D",
            "type": "object"
          },
          "preference_counts": {
            "additionalProperties": true,
            "title": "Preference Counts",
            "type": "object"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "device_id",
          "user_id",
          "created_at",
          "preference_counts",
          "engagement_7d"
        ],
        "title": "UserStatsResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI-powered tech news aggregator API. Fetches articles from RSS feeds and videos from YouTube, processes with AI summarization, and serves personalized feeds.",
    "title": "blips-ai-news",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/session/cache": {
      "delete": {
        "description": "Invalidate user's playlist cache (debug endpoint).",
        "operationId": "invalidate_cache_api_v1_session_cache_delete",
        "parameters": [
          {
            "description": "Device identifier",
            "in": "header",
            "name": "x-device-id",
            "required": true,
            "schema": {
              "description": "Device identifier",
              "title": "X-Device-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Invalidate Cache",
        "tags": [
          "session"
        ]
      }
    },
    "/api/v1/session/data": {
      "delete": {
        "description": "Delete all data associated with a device.\n\nRemoves usage records, interaction events, learned preferences, and the\ndevice profile.  Returns counts of deleted rows per table.",
        "operationId": "delete_my_data_api_v1_session_data_delete",
        "parameters": [
          {
            "description": "Device identifier",
            "in": "header",
            "name": "x-device-id",
            "required": true,
            "schema": {
              "description": "Device identifier",
              "title": "X-Device-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataDeletionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete My Data",
        "tags": [
          "session"
        ]
      }
    },
    "/api/v1/session/interactions": {
      "post": {
        "description": "Record a user interaction with content.\n\nInteraction types:\n- VIEW_10S: Viewed content for 10+ seconds\n- OPEN_SOURCE: Clicked to open source article/video\n- SHARE: Shared content\n- SAVE: Saved/bookmarked content\n- CHAT_START: Started AI chat about content\n- CHAT_MESSAGE: Sent message in AI chat\n\nUpdates user preferences based on interaction signals.",
        "operationId": "record_interaction_api_v1_session_interactions_post",
        "parameters": [
          {
            "description": "Device identifier",
            "in": "header",
            "name": "x-device-id",
            "required": true,
            "schema": {
              "description": "Device identifier",
              "title": "X-Device-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Interaction",
        "tags": [
          "session"
        ]
      }
    },
    "/api/v1/session/playlist": {
      "get": {
        "description": "Get a personalized playlist for the user session.\n\nUses session snapshots: once generated, a playlist is immutable\nfor that session. Use cursor-based continuation for swipe feeds.\n\nReturns ranked content items with diversity constraints:\n- No repeated stories (cluster_id)\n- Topic diversity (max 40% per topic)\n- Source rotation\n\nSession snapshot cached for 1 hour.",
        "operationId": "get_playlist_api_v1_session_playlist_get",
        "parameters": [
          {
            "description": "Content type",
            "in": "query",
            "name": "type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ContentTypeParam",
              "description": "Content type"
            }
          },
          {
            "description": "Items per request",
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Items per request",
              "maximum": 100,
              "minimum": 10,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "description": "Session ID for continuity",
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Session ID for continuity",
              "title": "Session Id"
            }
          },
          {
            "description": "Cursor position to continue from",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor position to continue from",
              "title": "Cursor"
            }
          },
          {
            "description": "Force refresh playlist (new session)",
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": false,
              "description": "Force refresh playlist (new session)",
              "title": "Refresh",
              "type": "boolean"
            }
          },
          {
            "description": "Device identifier",
            "in": "header",
            "name": "x-device-id",
            "required": true,
            "schema": {
              "description": "Device identifier",
              "title": "X-Device-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Playlist",
        "tags": [
          "session"
        ]
      }
    },
...
    "/api/v1/session/preferences": {
      "get": {
        "description": "Get user's current preferences (debug endpoint).\n\nReturns learned preferences across topics, entities, sources, and formats.",
        "operationId": "get_preferences_api_v1_session_preferences_get",
        "parameters": [
          {
            "description": "Device identifier",
            "in": "header",
            "name": "x-device-id",
            "required": true,
            "schema": {
              "description": "Device identifier",
              "title": "X-Device-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Preferences",
        "tags": [
          "session"
        ]
      }
    },
    "/api/v1/session/stats": {
      "get": {
        "description": "Get user statistics (debug endpoint).\n\nReturns preference counts and engagement metrics.",
        "operationId": "get_user_stats_api_v1_session_stats_get",
        "parameters": [
          {
            "description": "Device identifier",
            "in": "header",
            "name": "x-device-id",
            "required": true,
            "schema": {
              "description": "Device identifier",
              "title": "X-Device-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get User Stats",
        "tags": [
          "session"
        ]
      }
    },
...
  }
}